pyproject.toml
[project]
name = "doc2response"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "datagero", email = "datagero@outlook.com" }]
dependencies = [
    "llama-index>=0.11.20",
    "llama-index-vector-stores-faiss>=0.2.1",
    "faiss-cpu>=1.9.0",
    "python-dotenv>=1.0.1,<2",
    "pymysql>=1.1.1,<2",
    "llama-index-vector-stores-tidbvector>=0.2.0",
    "pypdf2>=3.0.1,<4",
    "pydantic>=2.5",
    "pydantic-settings>=2",
    "pydantic-ai>=0.0.14",
]

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "interfaces",
    "src",
]
